added ready application to get: you give start part,end part,email whom you want to read messages,it writes the part of the emails to the given file
